requirements.txt
pandas
numpy
scikit-learn
lightgbm
mlflow
joblib
tqdm
matplotlib
seaborn
pyyaml
wordcloud
streamlit
requests
prometheus-client
psutil
fastapi
uvicorn
ipykernel
arxivscraper

